Move fixtures to fixture folder
